Add check for flags
We can now check for the arguments for whether we should number
each line in the output. The number-nonblank argument has a higher
priority over the number argument. Depending on which argument is
used, we call different functions to output the texts and number
them if necessary. We used separate functions to avoid too many
nested if-statements. The code specifically emulates the GNU
version of cat which increases the line count with every line
across multiple files, as opposed to the BSD version which resets
the line count with each new file.